Remove unnecessary code Fix warnings Fix API url paths Add missing implementation to MNBService Add loggings Add some exception handling

+++ ExchangeRateBackend/ExchangeRateBackend/Services/Implementations/ExchangeRateService.cs
@@ -1,9 +1,10 @@
 ﻿using AutoMapper;
 using ExchangeRateBackend.Database;
 using ExchangeRateBackend.Models.Database;
-using ExchangeRateBackend.Models.RequestResponse;
+using ExchangeRateBackend.Models.Request;
 using ExchangeRateBackend.Models.Service;
 using ExchangeRateBackend.Services.Interfaces;
+using Microsoft.EntityFrameworkCore;
 
 namespace ExchangeRateBackend.Services.Implementations
 {
@@ -11,11 +12,13 @@ namespace ExchangeRateBackend.Services.Implementations
     {
         private readonly IMNBDataService _MNBDataService;
         private readonly IMapper _mapper;
+        private readonly ILogger<ExchangeRateService> _logger;
 
-        public ExchangeRateService(IMNBDataService MNBDataService, IMapper mapper)
+        public ExchangeRateService(IMNBDataService MNBDataService, IMapper mapper, ILogger<ExchangeRateService> logger)
         {
             _MNBDataService = MNBDataService;
             _mapper = mapper;
+            _logger = logger;
         }
 
         public async Task<List<ExchangeRate>> GetCurrentExchangeRatesAsync()
@@ -25,7 +28,7 @@ namespace ExchangeRateBackend.Services.Implementations
             return exchangeRates;
         }
 
-        public async Task<List<SavedExchangeRate>> GetSavedExchangeRatesAsync(int userId)
+        public List<SavedExchangeRate> GetSavedExchangeRates(int userId)
         {
             using (var db = new DatabaseContext())
             {
@@ -38,7 +41,7 @@ namespace ExchangeRateBackend.Services.Implementations
         {
             using (var db = new DatabaseContext())
             {
-                var model = db.SavedExchangeRates.FirstOrDefault(t => t.Id == id);
+                var model = await db.SavedExchangeRates.FirstOrDefaultAsync(t => t.Id == id);
                 return _mapper.Map<SavedExchangeRate>(model);
             }
         }
@@ -47,6 +50,12 @@ namespace ExchangeRateBackend.Services.Implementations
         {
             var exchangeRates = await GetCurrentExchangeRatesAsync();
             var currencyTo = exchangeRates.FirstOrDefault(t => t.Currency.Equals(to));
+            if(currencyTo == null)
+            {
+                var errorMessage = $"Currency not found: {to}";
+                _logger.LogError(errorMessage);
+                throw new Exception(errorMessage);
+            }
             var convertedValue = amount / currencyTo.Value;
             return convertedValue;
         }
@@ -64,17 +73,23 @@ namespace ExchangeRateBackend.Services.Implementations
                     UpdatedAt = DateTime.UtcNow,
                     UserId = data.UserId,
                 };
-                var model = db.SavedExchangeRates.Add(newData);
+                db.SavedExchangeRates.Add(newData);
                 await db.SaveChangesAsync();
                 return _mapper.Map<SavedExchangeRate>(newData);
             }
         }
 
-        public async Task<SavedExchangeRate> UpdateSavedExchangeRate(UpdateExchangeRateRequest data)
+        public async Task<SavedExchangeRate> UpdateSavedExchangeRateAsync(UpdateExchangeRateRequest data)
         {
             using (var db = new DatabaseContext())
             {
                 var model = db.SavedExchangeRates.FirstOrDefault(t => t.Id == data.Id);
+                if( model == null)
+                {
+                    var errorMessage = $"(UpdateSavedExchangeRate) Saved exchange rate with id not found: {data.Id}";
+                    _logger.LogError(errorMessage);
+                    throw new Exception(errorMessage);
+                }
                 model.UpdatedAt = DateTime.UtcNow;
                 model.Comment = data.Comment;
                 await db.SaveChangesAsync();
@@ -87,6 +102,12 @@ namespace ExchangeRateBackend.Services.Implementations
             using (var db = new DatabaseContext())
             {
                 var model = db.SavedExchangeRates.FirstOrDefault(t => t.Id == id);
+                if (model == null)
+                {
+                    var errorMessage = $"(DeleteSavedExchangeRate) Saved exchange rate with id not found: {id}";
+                    _logger.LogError(errorMessage);
+                    throw new Exception(errorMessage);
+                }
                 var asd = db.SavedExchangeRates.Remove(model);
                 await db.SaveChangesAsync();
                 return true;
